+++ src/Illuminate/Database/Eloquent/Builder.php
@@ -870,7 +870,11 @@ class Builder
             $values
         );
 
-        $values[$this->qualifyColumn($column)] = $values[$column];
+        $segments = preg_split('/\s+as\s+/i', $this->query->from);
+
+        $qualifiedColumn = end($segments).'.'.$column;
+
+        $values[$qualifiedColumn] = $values[$column];
 
         unset($values[$column]);
